Clean up controllers/user.ts by separating files
