+++ user/src/controllers/user.ts
@@ -1,39 +1,8 @@
-import { Request, Response, Router } from 'express';
-import { body, ValidationChain, validationResult } from 'express-validator';
+import { Request, Response } from 'express';
+import { validationResult } from 'express-validator';
 import { Document } from 'mongoose';
 import { createUser } from '../models/user';
-import { findUser, loginWithToken, registerUser, verifyUser } from '../models/userManager';
-
-const router: Router = Router();
-
-const userValidator: ValidationChain[] = [
-    body('email')
-        .not()
-        .isEmpty()
-        .withMessage('E-mail must be non-empty.')
-        .isEmail()
-        .withMessage('Must be a valid e-mail.'),
-    body('username')
-        .not()
-        .isEmpty()
-        .withMessage('Username must be non-empty.')
-        .isLength({ min: 6, max: 20 })
-        .withMessage('Username must be between 6 and 20 characters.'),
-    body('password')
-        .not()
-        .isEmpty()
-        .withMessage('Password must be non-empty.')
-        .isLength({ min: 6, max: 20 })
-        .withMessage('Password must be between 6 and 20 characters.'),
-    body('fullName')
-        .not()
-        .isEmpty()
-        .withMessage('Full name must be non-empty.')
-        .custom((value: string) => {
-            return /^[a-z ,.'-]+$/i.test(value);
-        })
-        .withMessage('Full name must be specified.'),
-];
+import { findUser, loginWithToken, registerUser } from '../models/userManager';
 
 const postUser = async (req: Request, res: Response): Promise<Response> => {
     const errors = validationResult(req);
@@ -53,26 +22,6 @@ const postUser = async (req: Request, res: Response): Promise<Response> => {
                 .status(201)
                 .set('token', token)
                 .send({ message: 'Successfully registered!' });
-        })
-        .catch((error: any) => {
-            return res.status(422).json({ error });
-        });
-};
-
-const postUserLogin = async (req: Request, res: Response): Promise<Response> => {
-    const { username, password } = req.body;
-    const token: any = req.header('token');
-
-    return verifyUser(username, password, token)
-        .then((newToken: string | false) => {
-            if (newToken) {
-                return res
-                    .status(200)
-                    .set('token', newToken)
-                    .send({ message: 'Successfully logged in.' });
-            } else {
-                return res.status(401).json({ message: 'Authorization denied.' });
-            }
         })
         .catch((error: any) => {
             return res.status(422).json({ error });
@@ -119,8 +68,4 @@ const getUser = async (req: Request, res: Response): Promise<Response> => {
     }
 };
 
-router.get('/user', getUser);
-router.post('/user', userValidator, postUser);
-router.post('/user/login', postUserLogin);
-
-export { router };
+export { getUser, postUser };
