+++ user/src/routes/routes.ts
@@ -1,4 +1,5 @@
 import { Application, Router } from 'express';
+import { postUserLogin } from '../controllers/login';
 
 export const initializeUserRoutes = (app: Application) => {
     const userRouter = Router();
@@ -6,21 +7,23 @@ export const initializeUserRoutes = (app: Application) => {
 
     /* gets a user */
     userRouter.get('/', async (req, res) => {
-        res.status(200).send("Working user get route")
+        res.status(200).send('Working user get route');
     });
 
     /* posts a user */
     userRouter.post('/', async (req, res) => {
-        res.status(200).send("Working user post route")
+        res.status(200).send('Working user post route');
     });
 
     /* deletes a user */
     userRouter.delete('/', async (req, res) => {
-        res.status(200).send("Working user delete route")
+        res.status(200).send('Working user delete route');
     });
 
     /* patches a user */
     userRouter.patch('/', async (req, res) => {
-        res.status(200).send("Working user patch route")
+        res.status(200).send('Working user patch route');
     });
+
+    userRouter.post('/login', postUserLogin);
 };
